+++ app/Http/Controllers/Builder/BuilderDownloadController.php
@@ -56,6 +56,8 @@ class BuilderDownloadController extends Controller
             }
         };
 
+        foreach ($items as $key => $value) if (isset($value['slots'])) $items[$key]['slots'] = $this->groupSlots($value['slots']);
+
         $data = [
             'name' => $inventory->name,
             'size' => $inventory->size,
@@ -68,6 +70,43 @@ class BuilderDownloadController extends Controller
             ->header('Content-Type', 'text/yaml')
             ->header("Content-Disposition", "attachment; filename={$inventory->file_name}.yml");
 
+    }
+
+    /**
+     * Permet de regrouper les slots qui se suivent pour réduire la configuration
+     *
+     * @param $tableau
+     * @return array
+     */
+    function groupSlots($tableau): array
+    {
+        sort($tableau);
+
+        $result = [];
+        $debut = $tableau[0];
+        $precedent = $debut;
+
+        for ($i = 1; $i < count($tableau); $i++) {
+            if ($tableau[$i] == $precedent + 1) {
+                $precedent = $tableau[$i];
+            } else {
+                if ($debut == $precedent) {
+                    $result[] = "$debut";
+                } else {
+                    $result[] = "$debut-$precedent";
+                }
+                $debut = $tableau[$i];
+                $precedent = $debut;
+            }
+        }
+
+        if ($debut == $precedent) {
+            $result[] = "$debut";
+        } else {
+            $result[] = "$debut-$precedent";
+        }
+
+        return $result;
     }
 
     public function is_in(InventoryButton $button, array $array): ?InventoryButton
@@ -91,17 +130,14 @@ class BuilderDownloadController extends Controller
             'slot' => $slot,
             'item' => [
                 'material' => $button->item->material,
-                'amount' => $button->amount,
             ]
         ];
 
-        if (isset($button->display_name) && $button->display_name !== null) {
-            $array['name'] = $button->display_name;
-        }
+        if ($button->amount != 1) $array['item']['amount'] = $button->amount;
 
-        if (isset($button->lore) && $button->lore !== null) {
-            $array['item']['lore'] = explode("\n", $button->lore);
-        }
+        if (isset($button->display_name)) $array['item']['name'] = $button->display_name;
+
+        if (isset($button->lore)) $array['item']['lore'] = explode("\n", $button->lore);
 
         if ($button->model_id != 0) $array['item']['modelId'] = $button->model_id;
         if ($button->glow) $array['item']['glow'] = true;
@@ -110,6 +146,12 @@ class BuilderDownloadController extends Controller
         if ($button->refresh_on_click) $array['refreshOnClick'] = true;
         if ($button->update_on_click) $array['updateOnClick'] = true;
         if ($button->update) $array['update'] = true;
+        if (isset($button->sound)) $array['sound'] = $button->sound;
+        if ($button->pitch != 1) $array['pitch'] = $button->pitch;
+        if ($button->volume != 1) $array['volume'] = $button->volume;
+        if (isset($button->messages)) $array['messages'] = explode("\n", $button->messages);
+        if (isset($button->commands)) $array['commands'] = explode("\n", $button->commands);
+        if (isset($button->console_commands)) $array['consoleCommands'] = explode("\n", $button->console_commands);
 
         return $array;
     }