+++ resources/js/builder/inventory/components/configuration/HeaderInformation.tsx
@@ -1,6 +1,11 @@
 import api from '../../../services/api'
+import LockScrollButton from "../utils/LockScrollButton";
+import {useState} from "react";
+import InformationModal from "../modals/InformationModal";
 
 const HeaderInformation = ({needToUpdate, saveData, inventoryId}) => {
+
+    const [showModal, setShowModal] = useState(false);
 
     return (
         <div className={"header-information"}>
@@ -14,15 +19,20 @@ const HeaderInformation = ({needToUpdate, saveData, inventoryId}) => {
                         needToUpdate ? (
                             <div className={'disable'}>
                                 <i className="bi bi-cloud-download"></i>
-                                <span className={"ms-1"}>Download the inventory</span>
+                                <span className={"ms-1"}>Download</span>
                             </div>
                         ) : (
                             <a className={'enable action'} href={api.getDownloadUrl(inventoryId)} target={"_blank"}>
                                 <i className="bi bi-cloud-download"></i>
-                                <span className={"ms-1"}>Download the inventory</span>
+                                <span className={"ms-1"}>Download</span>
                             </a>
                         )
                     }
+                    <LockScrollButton/>
+                    <span className={'action'} onClick={() => setShowModal(true)}>
+                        <i className="bi bi-info-lg"></i> Informations
+                    </span>
+                    <InformationModal handleClose={() => setShowModal(false)} show={showModal}/>
                 </div>
             </div>
         </div>