+++ resources/js/builder/inventory/components/configuration/itemstack/Glow.tsx
@@ -9,7 +9,7 @@ const Glow = ({currentSlot, handleChange}) => {
                 type="checkbox"
                 name="glow"
                 onChange={handleChange}
-                checked={currentSlot.button.glow}
+                checked={currentSlot.button?.glow ?? false}
                 className={'rounded-1'}
             />
             <a className={'ms-2'} href={'https://docs.zmenu.dev/configurations/items#glow'} target={'_blank'}>(<i className="bi bi-question-lg"></i>)</a>