+++ changelog.md
@@ -1,5 +1,16 @@
 # Unreleased
 
+# 1.0.4
+
+- Add informations modal
+- Add inventories element in user show page (admin)
+- Add folders page (admin)
+- Add messages, commands, console commands and sound configuration
+- Add a button to lock the page, to avoid scrolling and allow centered selections
+- Added grouping of slots when generating yaml file
+- Default slot selected will be slot 0
+- You can now move around the inventory with the directional keys
+
 # 1.0.3
 
 - Add multi slot support